feat
adcionando modularização para o modulo cliente
