+++ cliente.h
@@ -1,0 +1,9 @@
+
+//Assinaturas modulo clientes
+
+
+void tela_menu_cliente();
+    void cadast_cliente(void);
+    void exibir_cliente(void);
+    void edit_cliente(void);
+    void delet_cliente(void);
